### 2022-04-20 #4331 @editorialbot
[PRE REVIEW]: gofcat: An R package for goodness-of-fit of categorical response models
**Submitting author:** @ejikeugba (<a href="http://orcid.org/0000-0003-2572-0023">Ejike Ugba</a>)
**Repository:** https://github.com/ejikeugba/gofcat
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
