+++ joss/paper.bib
@@ -92,7 +92,6 @@
 	Journal = {Statistics in Medicine},
 	Volume = {23},
 	Year = {2004},
-	doi = {doi:10.1002/sim.1659},
 	Pages = {999--1014}
 }
 
@@ -132,7 +131,6 @@
 	Journal = {Mon. Weather. Rev.},
 	Volume = {78},
 	Year = {1950},
-	doi = {10.1175/1520-0493(1950)078%3C0001:VOFEIT%3E2.0.CO;2},
 	Pages = {1--3}
 }
 
@@ -142,6 +140,7 @@
 	Journal = {Biometrics},
 	Volume = {46},
 	Year = {1990},
+	doi = {10.2307/2532457},
 	Pages = {1171--1178}
 }
 
@@ -287,7 +286,6 @@
 	Journal = {Diabetic Med},
 	Volume = {13},
 	Year = {1996},
-	doi = {10.1002/(SICI)1096-9136(199606)13:6<536::AID-DIA110>3.0.CO;2-J},
 	Pages = {536--543}
 }
 
