### 2022-04-25 #4331 @danielskatz
👋 @sbenthall - would you be able to edit this submission for JOSS?
