+++ joss/paper.bib
@@ -92,6 +92,7 @@
 	Journal = {Statistics in Medicine},
 	Volume = {23},
 	Year = {2004},
+	doi = {10.1002/sim.1659},
 	Pages = {999--1014}
 }
 
@@ -131,6 +132,7 @@
 	Journal = {Mon. Weather. Rev.},
 	Volume = {78},
 	Year = {1950},
+	doi = {10.1175/1520-0493(1950)078<0001:vofeit>2.0.co;2},
 	Pages = {1--3}
 }
 
@@ -286,6 +288,7 @@
 	Journal = {Diabetic Med},
 	Volume = {13},
 	Year = {1996},
+	doi = {10.1002/(sici)1096-9136(199606)13:6<536::aid-dia110>3.0.co;2-j},
 	Pages = {536--543}
 }
 
