+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 2.7.3
-pkgdown: 2.0.2
+pkgdown: 2.0.3
 pkgdown_sha: ~
 articles: {}
-last_built: 2022-04-21T15:29Z
+last_built: 2022-04-25T18:12Z
 
