+++ joss/paper.md
@@ -68,7 +68,7 @@ Table: Proportional odds model of the retinopathy dataset. The common significan
 Applicable GOF tests for the fitted model (RET.fit) together with other yardsticks of model evaluation were further obtained using the available `gofcat` functions. A brief overview of what each function does is given, followed by an application to the model under consideration.
 
 ### hosmerlem() 
-This function performs the Hosmer-Lemeshow (HL) test for CRMs. Details of the test for the binary outcome model are given in @hosmer_goodness_1980. An extension to multinomial models is found in @fagerland_multinomial_2008 and @fagerland_generalized_2012, while the extension to ordinal models is found in Fagerland and Hosmer [-@fagerland_2013; -@fagerland_2016; -@fagerland_2017]. In each of the three settings, one splits the original observations into k groups (10 by default), after ranking by ordinal scores (OS). A Pearson Chi-square Statistic is then obtained from the expected and observed frequency tables. For a good chi-square approximation, it is recommended to have at least 80% of the estimated frequencies $>1$ [@fagerland_2017]. About 97% of the estimated frequencies from RET.fit meet this criterion. However, as observed in Table 2, the HL test is significant (p = 0.014), indicating a possible lack of fit.  
+This function performs the Hosmer-Lemeshow (HL) test for CRMs. Details of the test for the binary outcome model are given in @hosmer_goodness_1980. An extension to multinomial models is found in @fagerland_multinomial_2008 and @fagerland_generalized_2012, while the extension to ordinal models is found in Fagerland and Hosmer [-@fagerland_2013; -@fagerland_2016; -@fagerland_2017]. In each of the three settings, one splits the original observations into k groups (default to 10) after ranking by ordinal scores (OS). Ties are separated by an additional (within-group) ordering using the observed values. A Pearson Chi-square Statistic is then obtained from the expected and observed frequency tables. For a good chi-square approximation, it is recommended to have at least 80% of the estimated frequencies $>1$ [@fagerland_2017]. About 97% of the estimated frequencies from RET.fit meet this criterion. However, as observed in Table 2, the HL test is significant (p = 0.014), indicating a possible lack of fit.  
 
 ```r
 hosmerlem(RET.fit, tables = TRUE)
