### 2022-04-25 #4331 @ejikeugba
Thanks! @danielskatz. Some of the suggested DOIs still turned out invalid after having them included. I want to believe the DOIs are correct. Could it be that the editorialbot doesn't accept DOIs with semicolons? Because the reported invalid DOIs seem to have the semicolons initially in them missing.
