+++ joss/paper.md
@@ -68,21 +68,21 @@ Table: Proportional odds model of the retinopathy dataset. The common significan
 Applicable GOF tests for the fitted model (RET.fit) together with other yardsticks of model evaluation were further obtained using the available `gofcat` functions. A brief overview of what each function does is given, followed by an application to the model under consideration.
 
 ### hosmerlem() 
-This function performs the Hosmer-Lemeshow (HL) test for CRMs. Details of the test for the binary outcome model are given in @hosmer_goodness_1980. An extension to multinomial models is found in @fagerland_multinomial_2008 and @fagerland_generalized_2012, while the extension to ordinal models is found in Fagerland and Hosmer [-@fagerland_2013; -@fagerland_2016; -@fagerland_2017]. In each of the three settings, one splits the original observations into k groups (default to 10) after ranking by ordinal scores (OS). Ties are separated by an additional (within-group) ordering using the observed values. A Pearson Chi-square Statistic is then obtained from the expected and observed frequency tables. For a good chi-square approximation, it is recommended to have at least 80% of the estimated frequencies $>1$ [@fagerland_2017]. About 97% of the estimated frequencies from RET.fit meet this criterion. However, as observed in Table 2, the HL test is significant (p = 0.014), indicating a possible lack of fit.  
+This function performs the Hosmer-Lemeshow (HL) test for CRMs. Details of the test for the binary outcome model are given in @hosmer_goodness_1980. An extension to multinomial models is found in @fagerland_multinomial_2008 and @fagerland_generalized_2012, while the extension to ordinal models is found in Fagerland and Hosmer [-@fagerland_2013; -@fagerland_2016; -@fagerland_2017]. In each of the three settings, one splits the original observations into $k$ groups (default to 10) after ranking by ordinal scores and the response variable. The second odering criterion ensures that ties within the ordinal scores are separated. A Pearson chi-squared statistic is then obtained from the expected and observed frequency tables. For good a chi-squared approximation, it is recommended to have at least 80% of the estimated frequencies $>1$ [@fagerland_2017]. About 97% of the estimated frequencies from RET.fit meet this criterion. However, as observed in Table 2, the HL test is significant (p = 0.014), indicating a possible lack of fit.    
 
 ```r
 hosmerlem(RET.fit, tables = TRUE)
 ```
 
 ### lipsitz()
-This function computes the Lipsitz test for the ordinal models [@lipsitz_goodness_1996]. For this test, one also splits the observations into $k$ separate groups using the ordinal scores of the estimated values. An indicator variable denotes the observations belonging to each group, producing additional pseudo-variables with which the original model could be updated. Supposing the original model fits correctly, then the coefficients of the pseudo-variables all equal zero. The likelihood ratio statistic obtained from the log-likelihoods of the original and the refitted models is subsequently compared with the chi-square distribution having $k - 1$ degrees of freedom. In contrast to the LH test (Table 2), the Lipsitz test for the RET.fit is not statistically significant $(p = 0.459)$, implying that no lack of fit was detected. More discussion on this follows shortly.
[…]
