### 2022-04-25 #4331 @ejikeugba
@danielskatz sorry I didn't describe the problem well enough. The two invalid links returned by editorialbot are not the same as those in the references. These are the correct DOIs with the line numbers they appeared in the paper,

https://doi.org/10.1175/1520-0493(1950)078%3C0001:vofeit%3E2.0.co;2                     (line 159)
https://doi.org/10.1002/(sici)1096-9136(199606)13:6%3C536::aid-dia110%3E3.0.co;2-j      (line 209)

Both links work perfectly fine when clicked from the paper. However, editorialbot seems to reformat the link while running its checks, resulting in different DOIs altogether. I have no idea why it is so.
