### 2022-04-26 #4331 @danielskatz
👋 @openjournals/dev - can you check on this problem?  Perhaps there's a bug in how we're processing some strings/DOIs?
