### 2022-04-26 #4331 @xuanxu
Nice catch! yes, the `<` and `>` characters where not being escaped and so the generated url was invalid.
It will be fixed ASAP!
