### 2022-05-06 #4382 @editorialbot
[REVIEW]: gofcat: An R package for goodness-of-fit of categorical response models
**Submitting author:** @ejikeugba (<a href="http://orcid.org/0000-0003-2572-0023">Ejike Ugba</a>)
**Repository:** https://github.com/ejikeugba/gofcat
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.3
