### 2022-05-10 ejikeugba/gofcat#2 @NRaillard
`lipsitz()` behaviour when there is a `grp`variable
Hi,
Thanks for this package, I have a question while conducting some tests for the review ( openjournals/joss-reviews#4382 ).

What does happen if you have a variable called 'grp' in the data ? I suspect some results depending on the variable, see the snippet bellow, adapted from the README.

```r
dt <- data.frame(y = ordered(rbinom(n,2,0.5)), x1 = factor(rbinom(n,2,0.7)), x2 = runif(n))
dt$grp = dt$x1
sp <- serp(y ~ x1 +x2, slope="parallel", link = "logit", reverse= TRUE, data = dt)
sp_grp <- serp(y ~ grp +x2, slope="parallel", link = "logit", reverse= TRUE, data = dt)

#Same results
hosmerlem(sp,tables = TRUE)
hosmerlem(sp_grp,tables = TRUE)

#different results
lipsitz(sp)
lipsitz(sp_grp)
```
