### 2022-05-11 e95b9db
fix bugs in contr.fn
