### 2022-05-11 ejikeugba/gofcat#2 @ejikeugba
Hi @NRaillard!
Thanks for spotting it!  'grp' ends up being overwritten in the main function when used as a variable. A less likely used variable name now replaces it, also with a stopping message in case mistakenly used. This is fixed with [5ecea53](https://github.com/ejikeugba/gofcat/commit/5ecea53a14881d047aaa5399dd718c2952e86d68)
