### 2022-05-11 ejikeugba/gofcat#3 @ejikeugba
Thanks for spotting it! the bug is now fixed with [e95b9db](https://github.com/ejikeugba/gofcat/commit/e95b9db4a622ee71d55ed31f586f91294047ef9f)
