### 2022-05-16 #4382 @sbenthall
@PhilipPallmann Yes, that's appropriate! The checklist is a bare minimum of review. More thorough feedback and interaction with the author is even better!
