### 2022-07-08 #4382 @ejikeugba
@sbenthall , thanks for asking. It is not so clear yet why those points are left unchecked without corresponding comments or questions from @PhilipPallmann . It could be that he hasn't yet finished with his review. On my end, I believe the requirements for the unchecked points are all satisfied. I suggest you draw his attention to this. An update on how his review is going will be helpful. Thanks
