### 2022-08-18 #4382 @sbenthall
My supplemental review on the remaining items:
 *  Substantial scholarly effort: Commit history reveals several months of effort in the first five months of 2022 and end of 2021. The work builds on prior publications and has
 * Functionality: I'm not an R user, so it's hard for me to confirm this. But the other reviewer did check this off, and I see there are automated tests for the post-estimation tests that are the key contribution of the library.
 * Automated tests: There are automated tests covering many of the key post-estimation tests that are mentioned in the apper.
 * State of the field: The authors explain that there are many goodness-of-fit tests available in R, but not for categorical response models. Hence, the submission.
 * Quality of writing: It's fine writing.

As a supplemental editor, I recommend this article for acceptance to myself, @sbenthall , as editor.
