### 2022-08-18 #4382 @sbenthall
Thanks @sbenthall for the timely review!

Ok, @ejikeugba we can move on to the final steps on this now. Sorry about that long delay.

Those steps are:
 - Final proofreading and DOI checking on the paper...
 - You tag a release and put the release into an archive
 - You report the version number of the release and DOI in this thread.

Let me do one more proofreading pass. I encourage you to do the same because this will be the final version.
