### 2022-08-22 d2de21e
add standard.yaml
