### 2022-08-22 #4382 @ejikeugba
@sbenthall, thanks for making out time to work on the paper. I proofread it one last time, and everything looks fine. Just a minor adjustment in line 101 where I moved the parenthesis (Table 2) a bit further since it refers to 'PR test' rather than 'RET.fit'.
