### 2022-08-22 #4382 @sbenthall
Great, @ejikeugba . I recommend this article for acceptance to JOSS. Thanks for your patience, and sorry about the long delay.

Editors-in-chief, please not the irregularity in the reviewing. I stepped in with a supplemental review at the suggestion of @danielskatz in the JOSS Slack. I leave it to you to decided if the review suffices.
