### 2022-08-23 #4382 @arfon
> Editors-in-chief, please not the irregularity in the reviewing. I stepped in with a supplemental review at the suggestion of @danielskatz in the JOSS Slack. I leave it to you to decided if the review suffices.

Thanks for the heads-up @sbenthall. This looks good to me.
